Add max warning level configuration to faildelay check
This allows the user to configure a maximum permitted check level for
faildelay checks.

You can provide a default for all tasks, as well as per task
configurations that have precedence over the task default.

The configuration is stored in the site config.php under the config
setting $CFG->tool_heartbeat_tasks

This should be an array with a form like the following

$CFG->tool_heartbeat_tasks = [
    '*' => [
        'maxfaildelaylevel' => 'warning',
    ],
    '\core\task\delete_unconfirmed_users_task' => [
        'maxfaildelaylevel' => 'critical',
    ],
];